requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
astroid==2.4.2            # via pylint
attrs==20.1.0             # via -r requirements.in, pytest
bleach==3.1.5             # via readme-renderer
calf==0.3.2               # via -r requirements.in
certifi==2020.6.20        # via requests
cffi==1.14.2              # via cryptography
chardet==3.0.4            # via requests
colorama==0.4.3           # via twine
coverage==5.2.1           # via pytest-cov
cryptography==3.0         # via secretstorage
docutils==0.16            # via readme-renderer
idna==2.10                # via requests
importlib-metadata==1.7.0  # via keyring, pluggy, pytest, twine
iniconfig==1.0.1          # via pytest
isort==5.4.2              # via pylint
jeepney==0.4.3            # via keyring, secretstorage
keyring==21.3.1           # via twine
lazy-object-proxy==1.4.3  # via astroid
mccabe==0.6.1             # via pylint
more-itertools==8.4.0     # via pytest
mypy-extensions==0.4.3    # via mypy
mypy==0.782               # via -r requirements.in
packaging==20.4           # via bleach, pytest
pkginfo==1.5.0.1          # via twine
pluggy==0.13.1            # via pytest
py==1.9.0                 # via pytest
pycparser==2.20           # via cffi
pygments==2.6.1           # via readme-renderer
pylint==2.6.0             # via -r requirements.in
pyparsing==2.4.7          # via packaging
pytest-cov==2.10.1        # via -r requirements.in
pytest-mock==3.3.0        # via -r requirements.in
pytest==6.0.1             # via -r requirements.in, pytest-cov, pytest-mock
readme-renderer==26.0     # via twine
requests-toolbelt==0.9.1  # via twine
requests==2.24.0          # via requests-toolbelt, twine
rfc3986==1.4.0            # via twine
secretstorage==3.1.2      # via keyring
six==1.15.0               # via astroid, bleach, cryptography, packaging, readme-renderer
toml==0.10.1              # via pylint, pytest
tqdm==4.48.2              # via twine
twine==3.2.0              # via -r requirements.in
typed-ast==1.4.1          # via astroid, mypy
typing-extensions==3.7.4.2  # via mypy
urllib3==1.25.10          # via requests
webencodings==0.5.1       # via bleach
wheel==0.35.1             # via -r requirements.in
wrapt==1.12.1             # via astroid
zipp==3.1.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
